requirements.txt
torch>=2.0
torchvision>=0.15
numpy
matplotlib
opencv-python-headless
nuscenes-devkit
pyquaternion
scipy
